requirements.txt
# Core LangChain ecosystem
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.12
langchain-text-splitters==0.3.1

# .env file manipulation
python-dotenv==1.2.1

# Vector database
chromadb==0.5.5

# Local LLM + embeddings
langchain-ollama==0.3.10
ollama==0.3.3

# PDF & text handling
pypdf==5.1.0

# Optional — if you later want OpenAI API support
langchain-openai==0.2.2
openai==1.52.0

# Optional — if you want to use Hugging Face embeddings instead of Ollama
sentence-transformers==3.2.1
